Bot/requirements.txt
python-telegram-bot==20.7
groq
apscheduler
google-api-python-client
google-auth
python-dotenv
sqlalchemy
fastapi
uvicorn
